backend/requirements.txt
# Core LangChain
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-chroma
langchain-pinecone

# Embeddings
sentence-transformers
torch
transformers
huggingface-hub

# Vector stores
chromadb
pinecone                      # modern pinecone SDK (replaces pinecone-client)

# LLM
groq

# Data ingestion
requests
beautifulsoup4
feedparser
pandas
yfinance
youtube-transcript-api
pytube
lxml
html5lib

# FastAPI
fastapi
uvicorn
python-dotenv
pydantic

# Scheduler
schedule

# Utilities
numpy
tqdm